requirements.txt
streamlit>=1.32.0
numpy>=1.26.0
pillow>=10.0.0
scikit-learn>=1.5.0
joblib>=1.4.0
# Optional for trained model mode:
# tensorflow>=2.15.0
